refact: models and tests
