+++ spec/import_csv_to_database_spec.rb
@@ -7,9 +7,9 @@ describe 'Trying extract data from default file' do
     csv_rows = ImportFromCsv.csv_file('data.csv')
     ImportFromCsv.save_csv_file(csv_rows)
 
-    all_clients = SelectTable.all('clients')
-    all_doctors = SelectTable.all('doctors')
-    all_tests = SelectTable.all('tests')
+    all_clients = Client.all
+    all_doctors = Doctor.all
+    all_tests = Test.all
 
     expect(all_tests.count).to eq 3900
     expect(all_tests[0]['result_token']).to eq 'IQCZ17'
